python-services/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdf2image==1.16.3
PyMuPDF==1.23.9
Pillow==10.1.0
openai==1.3.8
google-generativeai==0.3.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.13.0
pydantic==2.5.0
typing-extensions==4.8.0